feat: add passing tests
